move 'both' tests to separate file
