+++ types/ramda/tsconfig.json
@@ -37,6 +37,7 @@
         "test/assoc-tests.ts",
         "test/assocPath-tests.ts",
         "test/binary-tests.ts",
-        "test/bind-tests.ts"
+        "test/bind-tests.ts",
+        "test/both-tests.ts"
     ]
 }
